FIX: collect and postp should handle INBOX instead of cur
