+++ src/util.rs
@@ -1,23 +1,8 @@
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::fs;
 use std::env;
-use dirs;
 use cmd_lib::*;
 use colored::Colorize;
-
-const DATA_BASE : &str = ".local/share/todor";
-const INBOX_NAME :&str  = "INBOX";
-
-pub fn get_inbox_file(dir: Option<String>, inbox: Option<String>) -> PathBuf {
-    let base_path = dir.map(PathBuf::from).unwrap_or_else(|| {
-        dirs::home_dir()
-            .expect("cannot get home directory")
-            .join(DATA_BASE)
-    });
-    fs::create_dir_all(&base_path).expect("Failed to create base directory");
-
-    base_path.join(inbox.unwrap_or(INBOX_NAME.to_string())).with_extension("md")
-}
 
 pub fn glance_all(inbox_path: &Path) {
